+++ src/Umbraco.Commerce.ProductFeeds.Core/Features/FeedSettings/Application/FindSettingParams.cs
@@ -1,4 +1,4 @@
-namespace Umbraco.Commerce.ProductFeeds.Core.FeedSettings.Application
+namespace Umbraco.Commerce.ProductFeeds.Core.Features.FeedSettings.Application
 {
     public class FindSettingParams
     {
