+++ src/Umbraco.Commerce.ProductFeeds.Core/Features/FeedSettings/Application/ProductFeedSettingWriteModelValidator.cs
@@ -1,5 +1,4 @@
 using FluentValidation;
-using Umbraco.Commerce.ProductFeeds.Core.FeedSettings.Application;
 
 namespace Umbraco.Commerce.ProductFeeds.Core.Features.FeedSettings.Application
 {
