+++ src/Umbraco.Commerce.ProductFeeds.Core/Features/PropertyValueExtractors/Application/IMultipleValuePropertyExtractor.cs
@@ -1,6 +1,6 @@
 using Umbraco.Cms.Core.Models.PublishedContent;
 
-namespace Umbraco.Commerce.ProductFeeds.Core.PropertyValueExtractors.Application
+namespace Umbraco.Commerce.ProductFeeds.Core.Features.PropertyValueExtractors.Application
 {
     /// <summary>
     /// Use to extract value of property that contains multiple value like MultipleMediaPicker.
