+++ src/Umbraco.Commerce.ProductFeeds.Startup/Initializations/IUmbracoCommerceBuilderExtensions.cs
@@ -7,11 +7,10 @@ using Umbraco.Commerce.Extensions;
 using Umbraco.Commerce.ProductFeeds.Core.Common.Constants;
 using Umbraco.Commerce.ProductFeeds.Core.Features.FeedGenerators.Application;
 using Umbraco.Commerce.ProductFeeds.Core.Features.FeedGenerators.Implementations;
+using Umbraco.Commerce.ProductFeeds.Core.Features.FeedSettings.Application;
+using Umbraco.Commerce.ProductFeeds.Core.Features.ProductQueries.Application;
+using Umbraco.Commerce.ProductFeeds.Core.Features.ProductQueries.Implementations;
 using Umbraco.Commerce.ProductFeeds.Core.Features.PropertyValueExtractors.Implementations;
-using Umbraco.Commerce.ProductFeeds.Core.FeedSettings.Application;
-using Umbraco.Commerce.ProductFeeds.Core.ProductQueries.Application;
-using Umbraco.Commerce.ProductFeeds.Core.ProductQueries.Implementations;
-using Umbraco.Commerce.ProductFeeds.Extensions;
 using Umbraco.Commerce.ProductFeeds.Infrastructure.DtoMappings;
 using Umbraco.Commerce.ProductFeeds.Infrastructure.Implementations;
 using Umbraco.Commerce.ProductFeeds.Infrastructure.Migrations;
