+++ src/Umbraco.Commerce.ProductFeeds.Web/Apis/Backoffice/Controllers/ProductFeedSettingController.cs
@@ -17,7 +17,6 @@ using Umbraco.Commerce.Extensions;
 using Umbraco.Commerce.ProductFeeds.Core.Common.Constants;
 using Umbraco.Commerce.ProductFeeds.Core.Features.FeedSettings.Application;
 using Umbraco.Commerce.ProductFeeds.Core.Features.PropertyValueExtractors.Implementations;
-using Umbraco.Commerce.ProductFeeds.Core.FeedSettings.Application;
 using Umbraco.Commerce.ProductFeeds.Web.Apis.Backoffice;
 using Umbraco.Commerce.ProductFeeds.Web.Apis.Backoffice.Controllers.Models;
